For consistency: rename Signup to Register
