+++ src/routes/Routes.tsx
@@ -4,9 +4,9 @@ import ContentAdmin from '../features/admin/components/ContentAdmin';
 import Upload from '../features/admin/components/Upload';
 import { User } from '../features/auth/actions';
 import AccountConfirmation from '../features/auth/components/AccountConfirmation';
-import LoginPage from '../features/auth/components/LoginPage';
 import PasswordChange from '../features/auth/components/PasswordChange';
-import Signup from '../features/auth/components/Signup';
+import Register from '../features/auth/components/Register';
+import SignIn from '../features/auth/components/SignIn';
 import Article from '../features/content/components/Article';
 import ArticleList from '../features/content/components/ArticleList';
 import Flashcards from '../features/content/components/Flashcards';
@@ -40,8 +40,8 @@ const Routes: React.FC = () => (
       <Route exact={true} path="/content" component={PublicationList} />
       <Route exact={true} path="/search" component={SearchPage} />
       <Route exact={true} path="/about" component={AboutPage} />
-      <Route exact={true} path="/login" component={LoginPage} />
-      <Route exact={true} path="/signup" component={Signup} />
+      <Route exact={true} path="/login" component={SignIn} />
+      <Route exact={true} path="/register" component={Register} />
       <Route
         exact={true}
         path="/confirmation/:token"
